float64 to float32

+++ Detour/DetourMath.go
@@ -2,10 +2,10 @@ package detour
 
 import "math"
 
-func DtMathFabsf(x float64) float64             { return math.Abs(x) }
-func DtMathSqrtf(x float64) float64             { return math.Sqrt(x) }
-func DtMathFloorf(x float64) float64            { return math.Floor(x) }
-func DtMathCeilf(x float64) float64             { return math.Ceil(x) }
-func DtMathCosf(x float64) float64              { return math.Cos(x) }
-func DtMathSinf(x float64) float64              { return math.Sin(x) }
-func DtMathAtan2f(y float64, x float64) float64 { return math.Atan2(y, x) }
+func DtMathFabsf(x float32) float32             { return float32(math.Abs(float64(x))) }
+func DtMathSqrtf(x float32) float32             { return float32(math.Sqrt(float64(x))) }
+func DtMathFloorf(x float32) float32            { return float32(math.Floor(float64(x))) }
+func DtMathCeilf(x float32) float32             { return float32(math.Ceil(float64(x))) }
+func DtMathCosf(x float32) float32              { return float32(math.Cos(float64(x))) }
+func DtMathSinf(x float32) float32              { return float32(math.Sin(float64(x))) }
+func DtMathAtan2f(y float32, x float32) float32 { return float32(math.Atan2(float64(y), float64(x))) }
